// admin/src/api/generated/schemas/adminBatchCreateBlacklistAddressItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminBatchCreateBlacklistAddressItem = {
  address?: string;
  network?: string;
  risk_level?: string;
  source?: string;
  reason?: string;
  monitor_status?: string;
};